refactor(mapper): extract the hydration/serving lane into serving/hydration-lane.ts (Phase 1 stage 1a)
Pure code motion of the contiguous tail (annotateGroundMeatName + lines
3731-7154: hydrateAndSelectServing, buildFdcResult, buildOffResult,
selectServing, gramsForServing, computeMacros and every volume/package
helper) into src/lib/mapping/serving/hydration-lane.ts. 3,460 moved lines
differ from the pre-extraction tree on exactly 22 lines, all relative
import paths gaining one directory of depth — verified mechanically.

The mapper keeps its 10 call sites untouched and re-exports the public
serving API (hydrateAndSelectServing, findOwnFdcVolumeServing,
candidateHasVolumeServing, isTightNameGroup, buildOffResult) so the two
eval instruments, the path-keyed jest.mock in fail-closed.test.ts and all
four value-importing test files resolve unchanged. Newly-shared private
helpers are exported via a bottom-of-file list, not inline keywords, so
the five COPIED_HELPERS bodies stay byte-identical.

Same-commit instrument updates (winner-diff would otherwise hard-throw):
- copiedHelperSource() reads the new file; helpers hash unchanged at
  8c1e518d30331976, caller hash unchanged at 2aa35e26b52161d1
- HASHED_SOURCE_PATHS gains src/lib/mapping/serving (gitDirtyHash reads
  only direct .ts children, the new dir would be a blind spot)
- doc-check claim stale-cache-delete-is-provenance-filtered repointed
  (2@2 holds on the new file)

Gates: tsc 0; jest 165/3394/1 skipped, zero test edits; next build green;
doc-check 108/109 (the 1 red is a pre-existing mobile log-length claim).
